backend/requirements.txt
# ── Core Flask ─────────────────────────────────────────────────────────────────
Flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1

# ── Database ───────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.30
# Uncomment for PostgreSQL (install separately: pip install psycopg2-binary)
# psycopg2-binary==2.9.9

# ── Auth & security ────────────────────────────────────────────────────────────
PyJWT==2.8.0
bcrypt==4.1.3

# ── Environment / config ───────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── File uploads ───────────────────────────────────────────────────────────────
Werkzeug==3.0.3

# ── Speech-to-text (choose one or both) ───────────────────────────────────────
# Option A – OpenAI Whisper (local, no API key, best accuracy)
# openai-whisper==20231117

# Option B – Google SpeechRecognition (lighter, requires internet for Google backend)
# SpeechRecognition==3.10.4

# ── AI chatbot ─────────────────────────────────────────────────────────────────
# Required only when OPENAI_API_KEY is set in .env
openai==1.30.1
